fetch new problems over open311
This enables the display of existing reports from the back end on FMS
if the body is configured to do this.

Reports will not be created if they are missing an id, a lat or a long,
if the lat/long is outside the area covered by the body, if there is
already a report with a matching id, or if we can't parse out the
request time.