Hide 'updates not sent' message on reports sent via Open311
The 'updates are not sent to the council' message was incorrectly being
shown on reports where updates would in fact be sent. This was affecting
non-Bromley/Stevenage reports on fms.com and all cobrands using Open311.

This commit moves the logic out of the template and adds the
Problem->updates_sent_to_body method which inspects the receiving body's
Open311 configuration to determine whether updates will be sent.

The duplication of the Lewisham/Oxfordshire logic between Problem.pm and
send-comments isn't ideal but hopefully there won't be any new Open311
bodies that only send and don't receive updates. If there are we'll have
to look at refactoring that list.