Show loading message on /around page load
The map and sidebar are empty on /around when the page loads, and JS is used
to pull in the relevant reports and populate the page. The previous message
indicated that there were no reports at all, but in fact they were being loaded
asynchronously. This commit shows a more helpful message whilst the pins are
being loaded.