Remove category_options_copy.
This variable existed to keep a copy of the categories before it was
overwritten by report_edit_location (if a bad location was provided)
calling the variable setup functions again. But really, the function
shouldn't be overwriting the stash, it's only testing the lookup and
so should restore the stash at its end.