Fix check for visible sub map links.
I don't think this check could ever have worked, because items are invisible
if their parent is set to `display: none`, so it would hide the links bar if
it was already hidden, and show it if it was already shown.

Use a js- class for the movement of the feed item into the sub map links.