Fix issues with new report pin, improve cursor.
The "coerce pin ID to integer" behaviour to fix a previous issue was
coercing undefined to NaN, which then caused other problems with the
new report pin, such as fading out when hovered, and being the wrong
size.

Also improve the cursor handling, showing the grab/ grabbing cursors
where available.