[Buckinghamshire] Improve the site_code lookup to fetch nearest asset
The previous approach of requesting a small bounding box from the
WFS server and using the first asset, if any, wasn’t sufficiently
tolerant of reports that were made more than 5M away from a road,
of which there were many. The problem with extending the bounding
box used to fetch assets from the server is that the order of the
returned assets didn’t have any bearing on their proximity to the
centre of the bounding box (i.e., the location of the report). As
an alternative approach this commit changes the method to instead
fetch all assets within a much wider bounding box surrounding the
report and iterates across the returned features in order to find
the closest point and uses the site_code value from that feature.