allow front end to not send to specific body
add a do_not_send parameter to reports which allows the front end to specify a
comma separated list of bodies that reports should not be sent to even if
there is a valid contact match. This enables asset layers to override backend
body processing, e.g. if a body only accepts reports that are on a road this
allows the front end to specify that.