Make it easier to prevent a form_detail_placeholder being printed
Previously, if you didn’t want a hint to appear under the main "details"
textarea on the new report form, you had to override the entire
`form_report.html` template, or leave the hint element in the markup but
hide it with CSS.

Now, you can set `form_detail_placeholder` to a falsey value, and the
template won’t output the hint element at all. It also amends the
`aria-describedby` attribute on the textarea so it doesn’t end up
referencing a hint element that doesn’t exist.