Vagrant: Use the mysociety/fixmystreet box by default
This sets the default box to be `mysociety/fixmystreet` and provides
an optional argument, `--base-box`, that can be used to override this.

When using the `mysociety/fixmystreet` box, the Perl modules are
prebuilt in `/usr/share/fixmystreet/local` and are bind-mounted into
`/home/vagrant/fixmystreet/local`. This ensures that these are
compatible with the guest machine and speeds up an initial launch.

The `mysociety/fixmystreet` box is also based on Debian Stretch so is
a closer match to our current production environment.