[Buckinghamshire] Question for on-road flytipping.
This adds a question to check whether a flytipping report concerns
flytipping on a road or not. Client-side, it is asked when a road is
selected and destination adjusted according to the answer; server side
it uses the answer to direct any both-tier submissions (from e.g.
non-JavaScript or phone app).