[Northamptonshire] use Alloy vector tiles for assets
Use the Vector tiles layers which Alloy provided which don't require
and API key or paging. This code is still a bit work in progress as it
makes a few assumptions:
 * the zoom level is hard coded
 * adding assets to layers assumes things are not async when they are
 * we are using itemId which is not the correct thing

This uses an event to notify that loading a request for a new set of
tiles has started and then uses counting to check we've fetched them all
and hence can get the layer to check for an asset.

It also uses start times to throw away requests for tiles in a previous
request if they come in after a new request has started.

This also doesn't have anything for putting labels on assets because
that information doesn't seem to be in the tiles.

A final point is that the tiles can contain multiple assets at the same
coőrdinates and we don't do anything about that at the moment so the
user will just get the last one added when they click.

This includes a dump of all the layers from alloy and then matches them
up to categories displayed on FixMyStreet