Expandable list item component
`_item_expandable.html` is an alternative to regular list items, where
you can display extra information on the first click on the item, and
then also, optionally, offer a further action, as a separate button.

Right now the expandable list items aren’t used anywhere, but we’ll
soon be using them in the duplicate report suggestion UI for council
staff and the public.