[Northamptonshire] make no asset and emergency code cooperate
The code to stop submission of an emergency category and the code to
stop submission of reports with no assets were conflicting. This makes
the emergency code take precedence over the asset code as there is no
point in telling the user they have not selected an asset and then
telling them to phone the council.