Switch all Norwegian tests to Swedish.
On Macs, the Norwegian locale is no_NO, whereas on Debian it is nb_NO:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=206474

This can cause issues running the Norwegian tests on Macs with certain
setups (e.g. on my machine a Homebrew installation of perl and gettext
fails, but MacPorts and system pass). As Swedish is sv_SE in both, the
easiest solution is to switch the tests to Swedish.