Set fixmystreet.bodies sooner.
There is a small bug where if you go to a two-tier location, pick a
category that is sent to one of the two ("These will be sent to" text
updates to only show that one), then click elsewhere on the map, the
"These will be sent to" text resets to show both bodies. This happens
because fixmystreet.bodies is reset after the text has been correctly
set by the category change() setup (called on pin move also).