Make sure anonymous reports are marked anon irrespective of permissions
There was a small bug where the 'default_to_body' permission would
override the 'report anonymously' button when staff users were adding
reports to the site. The result of this was that the name of the
anonymous user record would be shown. No harm done, as that user's name
would likely be set to "Anonymous user" in config, but it resulted in
reports pages showing the slightly odd wording along the lines of:

   Reported in the Bin bags category by Anonymous user at 14:57 today

This commit ensures the contributing_as_body flag is mutually exclusive
with the 'report_anonymously' request parameter.