[Westminster] Run stopper check after layer check.
The stopper check runs first, which could mean that if the asset layer
has not yet loaded, the form is disabled when it should not be (e.g.  on
a TfL road in a road category). Make sure that we always run the stopper
check after the TfL layer check, in case it changes something.