Make sure extras hidden when only_send is active.
The extras may be a confusing stopper message that only applies to the
'main' body, or it may be extra questions that are now not required.
Either way, we should hide this information, and we need to make sure
the server no longer checks if required extra questions have been
answered.