strip quote from contact group names
When an open311 category has multiple groups it puts them in the groups
element using CSV escaping. This means that group names can sometimes
feature double quotes. If we send these to the front end then putting
them into the optgroup name attribute breaks the HTML as if the group
name is in the database as `"group name"` you end up with

    <optgroup name="" group="" name="">

instead of

    <optgroup name="group name">

Hence, stripping the double quotes.