[IsleOfWight] limit filter category on national around page
Only display non triage categories on this because getting the category
expansion to work if we use the triage categories isn't in place and may
not be. This is because when we get the list of reports using
/around/ajax there's no way to know which body it is for so we don't
know how to expand a triage category to the list of 'child' categories.
So, just display the non triage ones for now.