Fix issue loading duplicates when not to be shown.
If the duplicates.js file is loaded, then its mechanisms come into play,
and it fetches duplicates and hides the reporting form if there are any.
So we only want this to happen in the same situations as it being loaded
elsewhere, on a relevant cobrand and not just after signing in.