[TfL] Server-side red route lookup for new report categories
This commit checks the RedRoutes WFS layer on tilma to determine if the
point at which a new report is being made is on a TfL red route.
The returned categories are then adjusted accordingly:

 - If on a red route, all TfL categories as well as borough categories
   specific to street cleaning are returned.
 - If not on a red route, all borough categories as well as TfL
   categories that don't require a red route are returned.
 - This category tweaking doesn't happen on the TfL cobrand, as the JS
   handles it by signposting users to fixmystreet.com for borough
   reports.

Doing the lookup server side means the app always shows the right
categories to the user and prevents them e.g. sending a borough
flytipping report to TfL.

Fixes https://github.com/mysociety/fixmystreet-commercial/issues/1716