Add per report OpenGraph images
Includes a bit of refactoring of PhotoSet::get_image_data to make
it easier to call subsequent methods on the photo object. We then
do this to get the width and height.

Also adds width/height attributes to FixMyStreet::ImageMagick
Attributes are updated every time the image is transformed and before
the as_blob data, which also undefs the image, is returned so it always
present for subsequent calls.

Fixes #2394.