Fix 'all reports' link for users with categories but no from_body
In the rare situation that a user has no from_body set but does have
categories set in their extra metadata, the 'All reports' link would
point to /my/inspector_redirect but this URL would 404 if the user
did not have a from_body. This commit updates the template to use the
same logic when generating the 'all reports' link, so users without
from_body will never be sent to /my/inspector_redirect.

Fixes #2815.